fix(durable): require in-process Snapshots and split workflow tests
inprocess.New panics without Snapshots, matching Temporal. SessionWorkflow
tests move to workflow_test.go.

Fixes #134
