+++ durable/inprocess/runtime_test.go
@@ -84,7 +84,7 @@ func waitEvents(t *testing.T, rt *Runtime, id durable.SessionID, sub durable.Sub
 func TestCreateSessionPromptCompletedMessage(t *testing.T) {
 	ctx := t.Context()
 	model := scriptedComplete("hello from agent")
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
@@ -120,7 +120,7 @@ func TestBindThenPromptReadsWorkspace(t *testing.T) {
 		InvokeFn: workspaceReadModel,
 	}
 	cat := newCatalog(t, model, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir)))})
-	rt := New(Config{Catalog: cat, Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: cat, Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -182,7 +182,7 @@ func TestPrompt_readSkillFromOpenSkills(t *testing.T) {
 	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{
 		OpenVFS:    vfs.Tree(vfs.At("work", builtins.Local(t.TempDir()))),
 		OpenSkills: vfs.Tree(vfs.At("skills", builtins.Local(pack))),
-	}), Projection: vfs.DirectProjection{}})
+	}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -216,7 +216,7 @@ func TestUnbindThenPromptWorkspaceGone(t *testing.T) {
 	model := &testkit.ScriptedModel{
 		InvokeFn: workspaceReadModel,
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: bindLocalDocs(dir)}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: bindLocalDocs(dir)}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -283,7 +283,7 @@ func TestAskUserChoiceYieldsThenResumeCompletes(t *testing.T) {
 			}
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -386,7 +386,7 @@ func TestPrompt_parallelBatchHitlRunsRemainder(t *testing.T) {
 			gate,
 			tacklr.NewTool(tacklr.ToolConfig{Name: "beta", Handler: slow("from-beta")}),
 		}},
-	}), Projection: vfs.DirectProjection{}})
+	}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -435,7 +435,7 @@ func TestCancelWhileRunningEndsSubscriptionThenNextPromptRuns(t *testing.T) {
 			<-ctx.Done()
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -505,7 +505,7 @@ func TestCancelWhileRunningEndsSubscriptionThenNextPromptRuns(t *testing.T) {
 
 func TestCloseDeletesSnapshotPromptNotFound(t *testing.T) {
 	ctx := t.Context()
-	rt := New(Config{Catalog: newCatalog(t, scriptedComplete("x"), durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, scriptedComplete("x"), durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -538,7 +538,7 @@ func TestTwoPromptsShareSnapshot(t *testing.T) {
 			ch <- tacklr.LLMResponseChunk{Type: tacklr.StreamEventMessage, Content: "ok", IsComplete: true}
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{
 		AgentID: "default",
 		State:   map[string]any{"user": "Ryan"},
@@ -594,7 +594,7 @@ func TestNewSessionDoesNotLoadPreviousSnapshot(t *testing.T) {
 			ch <- tacklr.LLMResponseChunk{Type: tacklr.StreamEventMessage, Content: "ok", IsComplete: true}
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id1, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -653,7 +653,7 @@ func TestPrompt_readMissingPathCorrection(t *testing.T) {
 			}
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir)))}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir)))}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -710,7 +710,7 @@ func TestPrompt_toolFailedServiceString(t *testing.T) {
 	}
 	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{
 		Options: tacklr.AgentOptions{Tools: []*tacklr.Tool{boom}},
-	}), Projection: vfs.DirectProjection{}})
+	}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -905,7 +905,7 @@ func TestSubscribeReplaysPastBuffer(t *testing.T) {
 func TestInferenceErrorPublishesStreamEventError(t *testing.T) {
 	ctx := t.Context()
 	model := &testkit.ScriptedModel{InvokeErr: errors.New("model down")}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -936,7 +936,7 @@ func TestCreateSessionMountsThenPromptTokensRemount(t *testing.T) {
 	if err := os.WriteFile(filepath.Join(dir, "hello.txt"), []byte("from-workspace"), 0o644); err != nil {
 		t.Fatal(err)
 	}
-	rt := New(Config{Catalog: newCatalog(t, &testkit.ScriptedModel{InvokeFn: workspaceReadModel}, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir)))}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, &testkit.ScriptedModel{InvokeFn: workspaceReadModel}, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir)))}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{
 		AgentID: "default",
 		Mounts: []durable.MountRecipe{{
@@ -978,7 +978,7 @@ func TestBadWorkspaceBindingFailsTurn(t *testing.T) {
 	ctx := t.Context()
 	missing := filepath.Join(t.TempDir(), "missing")
 	model := &testkit.ScriptedModel{InvokeFn: workspaceReadModel}
-	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(missing)))}), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(missing)))}), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -1026,7 +1026,7 @@ func TestChildSession_inheritsAndStatusStaysRunning(t *testing.T) {
 			}},
 		},
 	})
-	rt := New(Config{Catalog: cat, Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: cat, Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	parent, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
@@ -1102,7 +1102,7 @@ func TestSpawnWorkerNestedDriverCompletes(t *testing.T) {
 			}},
 		},
 	}
-	rt := New(Config{Catalog: newCatalog(t, model, spec), Projection: vfs.DirectProjection{}})
+	rt := New(Config{Catalog: newCatalog(t, model, spec), Snapshots: NewMemorySnapshot(), Projection: vfs.DirectProjection{}})
 	id, err := rt.CreateSession(ctx, durable.CreateSession{AgentID: "default"})
 	if err != nil {
 		t.Fatal(err)
